Add expiration time to local cache
Attempt to expire local cache after a specified time. This is a quick
and dirty solution to make it possible for asynchronous updates to not
permanently mess with document fetches.

Example:
1. Client A fetches a document
2. Client B, using a separate session, updates this document
3. Client A fetches the document again, but sees no change

Client A will never see a change until client A sends a futile update
request. Only after the update request fails will the client be able to
notice a change. Apply a passive fix that will automatically expire the
local object if enough time has passed, and client A can eventually
notice a change by fetching the document.